refactor: Update Message class to make staffId, id, and happinessId optional
This commit modifies the Message class in the message.ts file to make the staffId, id, and happinessId parameters optional. By changing these parameters to optional, we allow for more flexibility when creating instances of the Message class. This change is necessary to accommodate certain scenarios where these properties may not be available or needed.

Refactors:
- Made staffId, id, and happinessId parameters optional in the Message class

Fixes #123