refactor: Update ChatRoomRepository to use default values for Message class
This commit modifies the ChatRoomRepository file to update the instantiation of the Message class. The staffId, id, and happinessId parameters are now set to default values of null and RoleEnum.USER. This change allows for more flexibility when creating instances of the Message class and ensures consistent behavior across different scenarios.

Refactors:
- Updated ChatRoomRepository to use default values for Message class

Fixes #123